requirements.txt
Django==2.0.6
pytz==2018.4
wheel==0.24.0
Django==2.0.6
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
pytz==2018.4
wheel==0.24.0
Django==2.0.6
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
psycopg2==2.7.4
pytz==2018.4
wheel==0.24.0
